# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProcessMaker API",
        "description": "",
        "contact": {
            "email": "info@processmaker.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/1.0"
        }
    ],
    "paths": {
...
        "/processes": {
            "get": {
                "tags": [
                    "Processes"
                ],
                "summary": "Returns all processes that the user has access to",
                "operationId": "getProcesses",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/order_by"
                    },
                    {
                        "$ref": "#/components/parameters/order_direction"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/status"
                    },
                    {
                        "$ref": "#/components/parameters/include"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of processes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Get list Process",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Process"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/metadata"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Processes"
                ],
                "summary": "Save a new process",
                "operationId": "createProcess",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProcessEditable"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateNewProcess"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProcessEditable": {
                "properties": {
                    "process_category_id": {
                        "description": "Represents a business process definition.",
                        "type": "string",
                        "format": "id"
                    },
                    "name": {
                        "description": "Represents a business process definition.",
                        "type": "string"
                    },
                    "description": {
                        "description": "Represents a business process definition.",
                        "type": "string"
                    },
                    "status": {
                        "description": "Represents a business process definition.",
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "INACTIVE"
                        ]
                    },
                    "pause_timer_start": {
                        "description": "Represents a business process definition.",
                        "type": "integer"
                    },
                    "cancel_screen_id": {
                        "description": "Represents a business process definition.",
                        "type": "integer"
                    },
                    "has_timer_start_events": {
                        "description": "Represents a business process definition.",
                        "type": "boolean"
                    },
                    "start_events": {
                        "description": "Represents a business process definition.",
                        "type": "string",
                        "format": "json"
                    }
                },
                "type": "object"
            },
            "Process": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ProcessEditable"
                    },
                    {
                        "properties": {
                            "user_id": {
                                "description": "Represents a business process definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "id": {
                                "description": "Represents a business process definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "deleted_at": {
                                "description": "Represents a business process definition.",
                                "type": "string",
                                "format": "date-time"
                            },
                            "created_at": {
                                "description": "Represents a business process definition.",
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "description": "Represents a business process definition.",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "CreateNewProcess": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ProcessEditable"
                    },
                    {
                        "$ref": "#/components/schemas/Process"
                    },
                    {
                        "properties": {
                            "notifications": {
                                "description": "Represents a business process definition.",
                                "type": "array[]"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "metadata": {
                "properties": {
                    "filter": {
                        "type": "string"
                    },
                    "sort_by": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    },
                    "count": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    },
                    "current_page": {
                        "type": "integer"
                    },
                    "form": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "filter": {
                "name": "filter",
                "in": "query",
                "description": "Filter results by string. Searches Name, Description, and Status. Status must match exactly. Others can be a substring.",
                "schema": {
                    "type": "string"
                }
            },
            "order_by": {
                "name": "order_by",
                "in": "query",
                "description": "Field to order results by",
                "schema": {
                    "type": "string"
                }
            },
            "status": {
                "name": "status",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "active",
                    "enum": [
                        "active",
                        "inactive"
                    ]
                }
            },
            "order_direction": {
                "name": "order_direction",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "asc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "default": "10"
                }
            },
            "include": {
                "name": "include",
                "in": "query",
                "description": "Include data from related models in payload. Comma seperated list.",
                "schema": {
                    "type": "string",
                    "default": ""
                }
            },
...
        },
        "securitySchemes": {
            "pm_api_auth_code": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "scopes": "*"
                    }
                }
            },
            "pm_api_bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "pm_api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
    "security": [
        {
            "pm_api_token": []
        },
        {
            "pm_api_key": []
        },
        {
            "pm_api_auth_code": []
        }
    ]
}